refactor(openai): move prompt cache utilities to openai/ module
Spec 2.5 complete: Clean up request-utils.ts by moving OpenAI-specific
prompt cache utilities to the openai/ module.

Move from request-utils.ts to openai/request-handling.ts:
- extractPromptCacheKey: extract prompt cache key from request body
- hashPromptCacheKey: hash cache key for safe logging

Update imports:
- routes/chat.ts
- routes/responses.ts

request-utils.ts now contains only generic HTTP/request utilities:
- readCookieToken, parseJsonIfPossible, readSingleHeader
- escapeHtml, normalizeRequestedModel
- isTrustedLocalBridgeAddress, copyInjectedResponseHeaders
- SUPPORTED_V1_ENDPOINTS, SUPPORTED_NATIVE_OLLAMA_ENDPOINTS
- ChatCompletionRequest, WebSearchToolRequest interfaces

All 579 tests pass.
