+++ src/routes/responses.ts
@@ -1,11 +1,13 @@
 import type { FastifyInstance, FastifyReply } from "fastify";
 
 import {
-  extractPromptCacheKey,
-  hashPromptCacheKey,
   copyInjectedResponseHeaders,
 } from "../lib/request-utils.js";
-import { summarizeResponsesRequestBody } from "../lib/openai/index.js";
+import {
+  extractPromptCacheKey,
+  hashPromptCacheKey,
+  summarizeResponsesRequestBody,
+} from "../lib/openai/index.js";
 import { isRecord } from "../lib/provider-utils.js";
 import {
   resolvableConcreteModelIds,
